requirements.txt
crewai==0.1.32
instructor==0.4.7
langchain==0.1.1
langchain-community==0.0.13
openai==1.8.0
pre-commit==3.6.0
pur==7.3.1
pylint==3.0.3
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-randomly==3.15.0
python-dotenv==1.0.0
PyYaML==6.0.1
black==23.12.1
